chore: moved merge into specific translation utils
